Fix image normalization and always return float tensors
Updated image and ground truth processing in dataloaders to ensure outputs are always float tensors, regardless of value range. This prevents inconsistent return types and potential downstream errors. Version bumped to 0.1.0.53.

+++ gcpds_cv_pykit/baseline/dataloaders/dataloader.py
@@ -158,7 +158,8 @@ class Segmentation_Dataset(Dataset):
         if img.float().max() > 1.0:
             img = img.float() / 255.0
         else:
-            return img
+            img = img.float()
+        return img
 
     def process_mask(self, mask_paths: Sequence[Union[str, Path]]) -> torch.Tensor:
         """
